requirements.txt
Django
djangorestframework
django-celery-beat
celery
redis